requirements.txt
streamlit==1.8.1
pillow
Cython==0.29.28
rudalle
rudalle==0.0.1rc7
torch==1.9.0+cu111 torchvision==0.10.0+cu111 torchaudio==0.9.0 -f https://download.pytorch.org/whl/torch_stable.html
scipy==1.8.0